Separate concerns for youtube player vs. view page, didnt test this a lot
